+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog follow https://keepachangelog.com/ format.
 
 ## [Unreleased]
 
+*  `epy`: Add frozen dataclass support for `epy.ContextManager`
+
 ## [1.9.4] - 2024-09-03
 
 *   Return Python 3.10 support.